@@ -1,14 +1,13 @@
 {
   "name": "cra-template-retypetron",
-  "description": "Create React Typescript Tailwindcss Electron App.",
-  "version": "1.0.0",
+  "description": "Create React Typescript Electron App.",
+  "version": "1.1.0",
   "license": "MIT",
   "keywords": [
     "create-react-app",
     "react",
-    "electron",
     "typescript",
-    "tailwindcss",
+    "electron",
     "template"
   ],
   "author": "ngdduy17427",
@@ -31,22 +30,26 @@
   },
   "devDependencies": {
     "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@types/node": "^20.2.1",
-    "@types/react": "^18.2.6",
-    "@types/react-dom": "^18.2.4",
-    "@typescript-eslint/parser": "^5.62.0",
+    "@types/node": "^20.11.17",
+    "@types/react": "^18.2.55",
+    "@types/react-dom": "^18.2.19",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@typescript-eslint/parser": "^6.21.0",
+    "concurrently": "^8.2.2",
     "copyfiles": "^2.4.1",
-    "electron": "^28.0.0",
+    "electron": "^28.2.3",
     "electron-builder": "^24.9.1",
-    "eslint": "^8.55.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-prettier": "^5.1.3",
     "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-tailwindcss": "^3.14.2",
     "npm-run-all": "^4.1.5",
-    "prettier": "^3.1.1",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
     "react-scripts": "^5.0.1",
-    "sass": "^1.69.5",
-    "tailwindcss": "^3.3.3",
-    "typescript": "^5.0.4"
+    "sass": "^1.70.0",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.3.3"
   }
 }
