@@ -1,10 +1,10 @@
-import { contextBridge, ipcRenderer } from "electron";
+import { contextBridge, ipcRenderer, IpcRendererEvent } from "electron";
 
 interface IElectronAPI {
-  send: (channel: string, data?: any) => Promise<void>;
+  send: (channel: string, ...args: any[]) => Promise<void>;
   receive: (
     channel: string,
-    callback: (event: Electron.IpcRendererEvent, args: any[]) => void
+    callback: (event: IpcRendererEvent, ...args: any[]) => void
   ) => Promise<void>;
   removeAllListeners: (channel: string) => Promise<void>;
 }
@@ -18,12 +18,12 @@ declare global {
 const validChannels = ["HelloWorld"];
 
 contextBridge.exposeInMainWorld("electronAPI", {
-  send: (channel: string, data: any) => {
-    if (validChannels.includes(channel)) ipcRenderer.send(channel, data);
+  send: (channel: string, ...args: any[]) => {
+    if (validChannels.includes(channel)) ipcRenderer.send(channel, args);
   },
-  receive: (channel: string, func: any) => {
+  receive: (channel: string, callback: (event: IpcRendererEvent, ...args: any[]) => void) => {
     if (validChannels.includes(channel))
-      ipcRenderer.on(channel, (event, ...args) => func(event, ...args));
+      ipcRenderer.on(channel, (event, ...args) => callback(event, ...args));
   },
   removeAllListeners: (channel: string) => {
     if (validChannels.includes(channel)) ipcRenderer.removeAllListeners(channel);
